Add cra-vsr NodeNetworkConfig reconciler
Create a reconciler for cra-vsr agent.

Its job is to expose a Reconcile() API that will be called when a new
NodeNetworkConfig is detected in the cluster.

When called it fetchs the lastest NodeNetworkConfig and will try to apply
it using a cra manager. After that it updates the status of this
NodeNetworkConfig depending of the success of the cra manager.
In case of success, it checks node health and in case of failure reverts
to the previous working config.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>