Add cra-vsr layer3 setup
Define API to setup the layer3.

It parses NodeNetworkConfig FabricVRFs, LocalVRFs and ClusterVRF.
For each FabricVRF and LocalVRF, it create a new VRF.
It also creates a cluster and management VRF.

Algorithm try to keep the same table-id if VRF already exist in running
config. Otherwise a new one is generated is a specific range.

Only for each FabricVRF, a new bridge and vxlan is created in it.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>