Add cra-vsr BGP layer setup
Define API to setup the BGP layer.
This is the translation of the frr.conf.tpl into go code for cra-vsr.

It setups global configuration, that is prefix-lists and route-maps that
are always there. And setups BGP part and static routes for default,
cluster, management, all fabric and all local VRFs.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>