Finish cra-vsr manager
End cra-vsr manager by implementing ApplyConfiguration() API.

Here is the scenario:
- Replace the candidate config with the startup one.
- Build a VRouter object from the provided NodeNetworkConfig, calling
  layer2, layer3 and layerBGP setup() APIs.
- Merge this VRouter config into the candidate config. Some parts of the
  config are not merged but replace the current one, there are:
    / routing ...                        <--- type GlobalRouting
    / main foo routing ...               <--- type Routing
    / main foo l3vrf bar routing ...     <--- type Routing
- Commit the candidate config into the running config.

After some times without traffic the NETCONF server can close the
session in the back of the agent.
We detect it during the next NETCONF request, the API will failed with
EOF error. When that happens we re-open a session and retry to apply the
configuration.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>