Add cra-vsr types
Define types to generate 6WIND VSR configuration.
Will be used to build configuration and marshal it into valid NETCONF
xml. And to un-marshal received configuration in order to parse it.
Add API to Sort() the configuration, usefull to compare two
configurations and be sure that everything is ordered in the same way.

Use nemith/netconf MergeStrategy enumerator, so add dependency for it.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>