Add temporary agent-cra-vsr metrics proxy
The needed metrics are actually not available in cra-vsr.
One solution is to temporary deploy the hacking monitor deamon in the
cra-vsr, it will exposes http API to requests the metrics.

In order to be able to request this http API from the outside, the
agent-cra-vsr need to play the role of proxy. It will forwards requests
it received on http://X.X.X.X/cra/metrics to the hacking monitor running
into the cra-vsr.
For that purpose a new environment variable has been added CRA_METRICS_URL,
that should defines the addresses to reach the cra-monitor in the cra-vsr.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>