Create cra-vsr netconf generic Send() API
This should be the job of the netconf APIs to check if session is open()
and to re-open() session if the previous one has been closed.
Create a generic Send() API that wrap netconf.session.Call() in a secure
way, by checking presence of session and checking if the remote has
closed it.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>