Add cra-vsr VRouterConfig and VRouterState types
Define three VRouter types:
- VRouter: Generic vrouter that does not specify any specific XML tag
  container. This is the type that is use internaly.
- VRouterConfig: This is a VRouter which define its container XML tag to
  be <config>.
- VRouterState: This is a VRouter which define its container XML tag to
  be <state>.

This allows full flexibility, when we dont know the container tag of the
vrouter we can use VRouter, both config and state will be accepted.
If we marshal it we have to define the container tag and in this case
VRouterConfig or VRouterState should be used.

As netconf GetUnmarshal does not know into which VRouter it should
unmarshal the received data, the caller has now to provide the output
pointer.

Signed-off-by: Romain Delhomel <romain.delhomel@6wind.com>